manus_agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.2.2
ollama==0.3.3
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.9.2
